app/requirements.txt
streamlit==1.40.0
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
python-dotenv==1.0.1
